# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TeamCloud",
    "description": "API for working with a TeamCloud instance.",
    "contact": {
      "name": "TeamCloud Dev Team",
      "url": "https://github.com/microsoft/TeamCloud/issues/new",
      "email": "Markus.Heiliger@microsoft.com"
    },
    "license": {
      "name": "TeamCloud is licensed under the MIT License",
      "url": "https://github.com/microsoft/TeamCloud/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/orgs/{organizationId}/projects/{projectId}/templates/{id}": {
      "get": {
        "tags": [
          "ComponentTemplates"
        ],
        "summary": "Gets the Component Template.",
        "operationId": "GetComponentTemplate",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a Component Template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComponentTemplateDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Project Component Template with the provided id was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/orgs/{organizationId}/projects/{projectId}": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "Gets a Project.",
        "operationId": "GetProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a Project.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Project with the specified Name or ID was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "Deletes a Project.",
        "operationId": "DeleteProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Starts deleting the specified Project. Returns a StatusResult object that can be used to track progress of the long-running operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Project with the specified name or ID was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/orgs/{organizationId}/templates": {
      "get": {
        "tags": [
          "ProjectTemplates"
        ],
        "summary": "Gets all Project Templates.",
        "operationId": "GetProjectTemplates",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all Project Templates.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTemplateListDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "ProjectTemplates"
        ],
        "summary": "Creates a new Project Template.",
        "operationId": "CreateProjectTemplate",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectTemplateDefinition"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new Project Template was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTemplateDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "A Project Template already exists with the ID provided in the request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/orgs/{organizationId}/templates/{projectTemplateId}": {
      "get": {
        "tags": [
          "ProjectTemplates"
        ],
        "summary": "Gets a Project Template.",
        "operationId": "GetProjectTemplate",
        "parameters": [
          {
            "name": "projectTemplateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a ProjectTemplate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTemplateDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A ProjectTemplate with the projectTemplateId provided was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "ProjectTemplates"
        ],
        "summary": "Updates an existing Project Template.",
        "operationId": "UpdateProjectTemplate",
        "parameters": [
          {
            "name": "projectTemplateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectTemplate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The ProjectTemplate was updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTemplateDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Project Template with the ID provided in the request body could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "ProjectTemplates"
        ],
        "summary": "Deletes a Project Template.",
        "operationId": "DeleteProjectTemplate",
        "parameters": [
          {
            "name": "projectTemplateId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The ProjectTemplate was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectTemplateDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A ProjectTemplate with the projectTemplateId provided was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResultError": {
        "type": "object",
        "properties": {
          "code": {
            "enum": [
              "Unknown",
              "Failed",
              "Conflict",
              "NotFound",
              "ServerError",
              "ValidationError",
              "Unauthorized",
              "Forbidden"
            ],
            "type": "string"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StatusResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "stateMessage": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultError"
            },
            "nullable": true
          },
          "_trackingId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RepositoryReference": {
        "required": [
          "provider",
          "type",
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "baselUrl": {
            "type": "string",
            "nullable": true
          },
          "mountUrl": {
            "type": "string",
            "nullable": true
          },
          "ref": {
            "type": "string",
            "nullable": true
          },
          "provider": {
            "enum": [
              "Unknown",
              "GitHub",
              "DevOps"
            ],
            "type": "string"
          },
          "type": {
            "enum": [
              "Unknown",
              "Tag",
              "Branch",
              "Hash"
            ],
            "type": "string"
          },
          "organization": {
            "type": "string",
            "nullable": true
          },
          "repository": {
            "type": "string",
            "nullable": true
          },
          "project": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ComponentTemplate": {
        "required": [
          "id",
          "organization",
          "parentId",
          "repository",
          "type"
        ],
        "type": "object",
        "properties": {
          "organization": {
            "type": "string"
          },
          "parentId": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "repository": {
            "$ref": "#/components/schemas/RepositoryReference"
          },
          "permissions": {
            "type": "object",
            "properties": {
              "None": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "Member": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "Admin": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "Owner": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "nullable": true
          },
          "inputJsonSchema": {
            "type": "string",
            "nullable": true
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentTaskTemplate"
            },
            "nullable": true
          },
          "taskRunner": {
            "$ref": "#/components/schemas/ComponentTaskRunner"
          },
          "type": {
            "enum": [
              "Custom",
              "AzureResource",
              "Environment",
              "GitRepository"
            ],
            "type": "string"
          },
          "folder": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ComponentTemplateDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ComponentTemplate"
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Project": {
        "required": [
          "displayName",
          "id",
          "organization",
          "slug",
          "template"
        ],
        "type": "object",
        "properties": {
          "organization": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "template": {
            "type": "string"
          },
          "templateInput": {
            "type": "string",
            "nullable": true
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "nullable": true
          },
          "tags": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "resourceState": {
            "enum": [
              "Pending",
              "Initializing",
              "Provisioning",
              "Succeeded",
              "Failed"
            ],
            "type": "string"
          },
          "vaultId": {
            "type": "string",
            "nullable": true
          },
          "storageId": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ProjectDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/Project"
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProjectTemplate": {
        "required": [
          "displayName",
          "id",
          "isDefault",
          "organization",
          "repository",
          "slug"
        ],
        "type": "object",
        "properties": {
          "organization": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string"
          },
          "components": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "repository": {
            "$ref": "#/components/schemas/RepositoryReference"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isDefault": {
            "type": "boolean"
          },
          "inputJsonSchema": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ProjectTemplateListDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectTemplate"
            },
            "nullable": true,
            "readOnly": true
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RepositoryDefinition": {
        "required": [
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProjectTemplateDefinition": {
        "required": [
          "displayName",
          "repository"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "repository": {
            "$ref": "#/components/schemas/RepositoryDefinition"
          }
        },
        "additionalProperties": false
      },
      "ProjectTemplateDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/ProjectTemplate"
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://login.microsoftonline.com//oauth2/v2.0/authorize",
            "tokenUrl": "https://login.microsoftonline.com//oauth2/v2.0/token",
            "scopes": {
              "openid": "Sign you in",
              "http://TeamCloud.aztcclitestsix/user_impersonation": "Access the TeamCloud API"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "openid",
        "http://TeamCloud.aztcclitestsix/user_impersonation"
      ]
    }
  ]
}